+++ changelog.md
@@ -58,3 +58,8 @@
     * finalized get_sessions
     * added get_session
     * update get_machines with pagination and filtering
+14-02-2021
+* Added Dev branch 
+* Inventory Class
+    * started work on delete_machines
+    * removed print that was left behind in get_machines